feat: add ios 15 and Android 12 support
