+++ package.json
@@ -32,31 +32,31 @@
   "author": "wswebcreation",
   "license": "MIT",
   "devDependencies": {
-    "@babel/cli": "^7.15.4",
-    "@babel/core": "^7.15.4",
-    "@babel/preset-env": "^7.15.4",
-    "@babel/register": "^7.15.3",
-    "@wdio/cli": "^7.11.1",
-    "@wdio/devtools-service": "^7.11.1",
-    "@wdio/jasmine-framework": "^7.11.1",
-    "@wdio/local-runner": "^7.11.1",
-    "@wdio/sauce-service": "^7.11.1",
-    "@wdio/selenium-standalone-service": "^7.10.1",
-    "@wdio/spec-reporter": "^7.10.1",
+    "@babel/cli": "^7.16.0",
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/register": "^7.16.0",
+    "@wdio/cli": "^7.16.9",
+    "@wdio/devtools-service": "^7.16.9",
+    "@wdio/jasmine-framework": "^7.16.6",
+    "@wdio/local-runner": "^7.16.9",
+    "@wdio/sauce-service": "^7.16.9",
+    "@wdio/selenium-standalone-service": "^7.16.6",
+    "@wdio/spec-reporter": "^7.16.9",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^27.1.0",
-    "devtools": "^7.11.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.24.2",
+    "babel-jest": "^27.3.1",
+    "devtools": "^7.16.7",
+    "eslint": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
     "fs-extra": "^10.0.0",
-    "jest": "^27.1.0",
-    "np": "^7.5.0",
+    "jest": "^27.3.1",
+    "np": "^7.6.0",
     "rimraf": "^3.0.2",
-    "webdriverio": "^7.11.1"
+    "webdriverio": "^7.16.9"
   },
   "dependencies": {
-    "@wdio/logger": "^7.7.0",
-    "webdriver-image-comparison": "0.19.0"
+    "@wdio/logger": "^7.16.0",
+    "webdriver-image-comparison": "0.20.0"
   },
   "repository": {
     "type": "git",
