+++ tests/sauceLabsBaseline/desktop_safari/tabbable-MojaveSafari12-1366x768.png
